+++ tests/test_containerproperties.py
@@ -1,9 +1,19 @@
+import dataclasses
 import datetime
 import unittest
+from unittest import mock
+
+from lxml import etree
 
 import sdc11073.mdib.containerproperties as containerproperties
+from sdc11073.definitions_sdc import SDC_v1_Definitions
+from sdc11073.mdib import clientmdib
+from sdc11073.mdib import devicemdib
+from sdc11073.mdib import msgreader
 from sdc11073.mdib.descriptorcontainers import ClockDescriptorContainer
-from sdc11073.namespaces import DocNamespaceHelper, domTag
+from sdc11073.namespaces import DocNamespaceHelper
+from sdc11073.namespaces import domTag
+from sdc11073.namespaces import msgTag
 
 # pylint: disable=protected-access
 
@@ -99,6 +109,105 @@ class TestContainerproperties(unittest.TestCase):
                                                 parentHandle='467')
         node2 = dc2.mkNode()
         self.assertEqual(node2.attrib['Resolution'], 'PT77S')
+
+    def test_activate_operation_argument(self):
+        """Test that pm:ActivateOperationDescriptor/pm:argument/pm:Arg is handled correctly
+        because its value is a QName"""
+        @dataclasses.dataclass
+        class TestData:
+            mdib_text: str
+            expected_qname: etree.QName
+
+        mdib_dummy = """<msg:GetMdibResponse 
+                        xmlns:msg="http://standards.ieee.org/downloads/11073/11073-10207-2017/message"
+                         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                         {0}
+                         xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
+                         xmlns:xsd="http://www.w3.org/2001/XMLSchema"
+                         xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
+                         MdibVersion="174" SequenceId="urn:uuid:6f4ff7de-6809-4883-9938-54dd4f6f9173">
+                        <msg:Mdib MdibVersion="174" SequenceId="urn:uuid:6f4ff7de-6809-4883-9938-54dd4f6f9173">
+                            <pm:MdDescription DescriptionVersion="8">
+                                <pm:Mds Handle="mds0" DescriptorVersion="5" SafetyClassification="MedA">
+                                    <pm:Sco Handle="Sco.mds0" DescriptorVersion="2">
+                                        <pm:Operation {1} xsi:type="pm:ActivateOperationDescriptor" Handle="SVO.38.3569" DescriptorVersion="2"
+                                                      SafetyClassification="MedA" OperationTarget="3569" MaxTimeToFinish="PT00H00M02S"
+                                                      Retriggerable="false">
+                                            <pm:Type Code="193821">
+                                                <pm:ConceptDescription Lang="en-US">An operation to cancel global all audio pause
+                                                </pm:ConceptDescription>
+                                            </pm:Type>
+                                            <pm:Argument>
+                                                <pm:ArgName Code="codeForArgumentName"></pm:ArgName>
+                                                <pm:Arg {2}>{3}duration</pm:Arg>
+                                            </pm:Argument>
+                                        </pm:Operation>
+                                    </pm:Sco>
+                                    <pm:SystemContext Handle="SC.mds0" DescriptorVersion="2">
+                                        <ext:Extension>
+                                        </ext:Extension>
+                                        <pm:PatientContext Handle="PC.mds0" DescriptorVersion="2">
+                                        </pm:PatientContext>
+                                        <pm:LocationContext Handle="LC.mds0" DescriptorVersion="2">
+                                        </pm:LocationContext>
+                                    </pm:SystemContext>
+                                </pm:Mds>
+                            </pm:MdDescription>
+                        </msg:Mdib>
+                    </msg:GetMdibResponse>"""
+
+        my_prefix = "my"
+        xsd_prefix = "xsd"
+        delaration = 'xmlns:{0}="http://www.w3.org/2001/XMLSchema"'
+        delaration_any_uri = 'xmlns:{0}="urn:oid:1.23.3.123.2"'
+        expected_qname_xsd = etree.QName("http://www.w3.org/2001/XMLSchema", "duration")
+        expected_qname_any_uri = etree.QName("urn:oid:1.23.3.123.2", "duration")
+
+        mdibs = [TestData(mdib_text=mdib_dummy.format('', '', delaration.format(my_prefix), f"{my_prefix}:"),
+                          expected_qname=expected_qname_xsd),
+                 TestData(mdib_text=mdib_dummy.format('', delaration.format(my_prefix), '', f"{my_prefix}:"),
+                          expected_qname=expected_qname_xsd),
+                 TestData(mdib_text=mdib_dummy.format(delaration.format(my_prefix), '', '', f"{my_prefix}:"),
+                          expected_qname=expected_qname_xsd),
+                 TestData(mdib_text=mdib_dummy.format('', '', delaration.format(xsd_prefix), f"{xsd_prefix}:"),
+                          expected_qname=expected_qname_xsd),
+                 TestData(mdib_text=mdib_dummy.format('', '', 'xmlns="http://www.w3.org/2001/XMLSchema"', ''),
+                          expected_qname=expected_qname_xsd),
+                 TestData(mdib_text=mdib_dummy.format('', '', delaration_any_uri.format(xsd_prefix), f"{xsd_prefix}:"),
+                          expected_qname=expected_qname_any_uri),
+                 TestData(mdib_text=mdib_dummy.format('', delaration_any_uri.format(xsd_prefix), '', f"{xsd_prefix}:"),
+                          expected_qname=expected_qname_any_uri)]
+
+        for test_data in mdibs:
+            # parse mdib data into container and reconstruct mdib data back to a msg:GetMdibResponse
+            # so that it can be validated by xml schema validator
+            mdib_text = test_data.mdib_text.encode('utf-8')
+            mdib_container = devicemdib.DeviceMdibContainer.fromString(mdib_text)
+            mdib_node, mdib_version_group = mdib_container.reconstructMdibWithContextStates()
+            getMdibResponseNode = etree.Element(msgTag('GetMdibResponse'), nsmap=mdib_node.nsmap)
+            mdib_version_group.update_node(getMdibResponseNode)
+            getMdibResponseNode.append(mdib_node)
+            mdib_container.sdc_definitions.xml_validator.assertValid(getMdibResponseNode)
+
+            # read msg:GetMdibResponse as an sdc11073 client would do
+            # and verify that pm:Argument/pm:Arg data is still valid
+
+            # Preparation
+            sdc_client = mock.Mock()
+            sdc_client.sdc_definitions = SDC_v1_Definitions
+            sdc_client.log_prefix = ""
+            client_mdib = clientmdib.ClientMdibContainer(sdc_client)
+            tmp_msg_reader = msgreader.MessageReader(client_mdib)
+
+            # parse getMdibResponseNode to container
+            descriptors = tmp_msg_reader.readMdDescription(getMdibResponseNode)
+            op_descr = [descr for descr in descriptors if descr.isOperationalDescriptor][0]
+            tmp_arg = op_descr.Argument[0].Arg
+            self.assertEqual(tmp_arg.localname, test_data.expected_qname.localname)
+            self.assertEqual(tmp_arg.namespace, test_data.expected_qname.namespace)
+            self.assertEqual(tmp_arg.text, test_data.expected_qname.text)
+            self.assertIn(tmp_arg.namespace, tmp_arg.nsmap.values())
+
 
 def suite():
     return unittest.TestLoader().loadTestsFromTestCase(TestContainerproperties)